Bugfix for error.stack not being accessible
When accessing error.stack from within WebKit, it returns a 'DOM
secutiry exception 18' because the URL is about:blank.
Therefore Phantomic now creates a web server on localhost to ship the
resources to WebKit to work around this problem.

Also fixed an issue with the first console.error call stopping script
execution.